ai_service/requirements.txt
fastapi
uvicorn[standard]
google-genai
motor
python-dotenv
hypothesis
pytest
pytest-asyncio
